backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic>=2.0.0
pydantic-settings>=2.0.0
supabase==2.10.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
celery[redis]==5.4.0
redis==5.2.1
openai==1.54.0
httpx==0.27.2
Pillow==11.0.0
pdf2image==1.17.0
flower==2.0.1

